PostgreSQL OIDs overflow fix
Former-commit-id: 31444098d91c2d6621ef653d7e7f932eca5c66f8

+++ plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/model/PostgreCharset.java
@@ -55,7 +55,7 @@ public class PostgreCharset extends PostgreInformation {
     }
 
     @Override
-    public int getObjectId() {
+    public long getObjectId() {
         return encodingId;
     }
